Wait period should be a float instead of int

+++ app/factionstats/settings.py
@@ -141,7 +141,7 @@ CORS_ORIGIN_WHITELIST = [
 
 CELERY_RESULT_BACKEND = 'django-db'
 
-TORN_API_RATE = int(os.environ.get('API_RATE', default=1))  # xx/yy Maximum of yy requests per xx seconds
+TORN_API_RATE = float(os.environ.get('API_RATE', default=1))  # xx/yy Maximum of yy requests per xx seconds
 TORN_API_BASE_URL = """https://api.torn.com/"""
 TORN_API_MIN_STATUS_DWELL_MINUTES = int(os.environ.get('API_MIN_STATUS_DWELL_MINUTES', default=5))
 TORN_API_MAX_TARGET_RETURN = int(os.environ.get('API_MAX_TARGET_RETURN', default=30))
